+++ src/sources/vector-table-source.ts
@@ -30,17 +30,22 @@ type UrlParameters = {
   spatialDataColumn?: string;
   tileResolution?: string;
   name: string;
+  aggregationExp?: string;
 };
+
+export type VectorTableSourceResponse = TilejsonResult &
+  WidgetTableSourceResult;
 
 export const vectorTableSource = async function (
   options: VectorTableSourceOptions
-): Promise<TilejsonResult & WidgetTableSourceResult> {
+): Promise<VectorTableSourceResponse> {
   const {
     columns,
     filters,
     spatialDataColumn = 'geom',
     tableName,
     tileResolution = DEFAULT_TILE_RESOLUTION,
+    aggregationExp,
   } = options;
 
   const urlParameters: UrlParameters = {
@@ -55,6 +60,9 @@ export const vectorTableSource = async function (
   }
   if (filters) {
     urlParameters.filters = filters;
+  }
+  if (aggregationExp) {
+    urlParameters.aggregationExp = aggregationExp;
   }
   return baseSource<UrlParameters>('table', options, urlParameters).then(
     (result) => ({